requirements.txt
bcrypt==3.2.0
email-validator==2.1.0.post1
fastapi==0.109.2
PyJWT==1.7.1
python-decouple==3.3
python-multipart==0.0.9
uvicorn==0.27.1